[update]色々変更

+++ Assets/Scripts/Runtime/1.Domain/MinionEntity.cs
@@ -2,12 +2,12 @@ namespace Domain
 {
     public class MinionEntity
     {
-        public MinionEntity(int health, 
-                            int defence, 
+        public MinionEntity(float health,
+                            float defence,
                             float moveSpeed,
-                            int attackPower, 
+                            float attackPower,
                             float attackRange,
-                            float attackSpeed ,
+                            float attackSpeed,
                             float criticalChance,
                             float criticalDamage)
         {
